Move Unit.Common functions into Scale, Format
Instead of placing them in their own "Common" modules, the functions are
now just part of the Scale and Format modules. To use the (formerly)
common "scale" function, instead of calling Benchee.Unit.Common.scale/1,
now you call Benchee.Conversion.Scale.scale/1. Likewise with Format
functions
