+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ## Breaking Changes (Plugins)
 
-* The `Benchee.Time`module is gone, if you relied on it for one reason or another it's succeeded by the more powerful `Benchee.Unit.Duration`
+* The `Benchee.Time`module is gone, if you relied on it for one reason or another it's succeeded by the more powerful `Benchee.Conversion.Duration`
 
 # 0.4.0 (September 11, 2016)
